+++ elasticsearch5/search-index/src/main/java/org/vertexium/elasticsearch5/ElasticsearchSearchQueryBase.java
@@ -403,7 +403,7 @@ public class ElasticsearchSearchQueryBase extends QueryBase {
                 if (propertyDefinition.getDataType() == String.class) {
                     scriptSrc += "return fieldValues.length > 0 ? fieldValues : (params.esOrder == 'asc' ? [Character.toString(Character.MAX_VALUE)] : '');";
                 } else {
-                    scriptSrc += "return fieldValues.length > 0 ? fieldValues[0] : (params.esOrder == 'asc' ? Integer.MAX_VALUE : Integer.MIN_VALUE);";
+                    scriptSrc += "return fieldValues.length > 0 ? fieldValues[0] : (params.esOrder == 'asc' ? Long.MAX_VALUE : Long.MIN_VALUE);";
                 }
 
                 List<String> fieldNames = Arrays.stream(propertyNames).map(propertyName ->